requirements_ccxt.txt
ccxt==4.2.25
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
